refactored all the code into a much more coherent, namespaces and libraries
