+++ apps/CMakeLists.txt
@@ -1,2 +1,10 @@
 add_executable(my_game main.cpp)
-target_link_libraries(my_game PRIVATE glad glfw window shaders stb_image shape game_obj glm game) 
+target_link_libraries(my_game PRIVATE
+    opengl_core
+    opengl_graphics  
+    opengl_geometry
+    opengl_game
+    glad
+    glfw
+    glm
+)
