+++ libs/internal_libs/CMakeLists.txt
@@ -1,5 +1,1 @@
-add_subdirectory(shaders)
-add_subdirectory(window)
-add_subdirectory(shape)
-add_subdirectory(game_obj)
-add_subdirectory(game)
+add_subdirectory(opengl)
