+++ libs/internal_libs/opengl/CMakeLists.txt
@@ -1,0 +1,4 @@
+add_subdirectory(core)
+add_subdirectory(game)
+add_subdirectory(geometry)
+add_subdirectory(graphics)
