+++ libs/internal_libs/opengl/core/include/shaders.h
@@ -1,0 +1,29 @@
+#ifndef SHADER_H
+#define SHADER_H
+#include <glad/glad.h>  // include glad to get the required OpenGL headers
+#include <glm/glm.hpp>
+#include <glm/gtc/type_ptr.hpp>
+#include <string>
+#include <fstream>
+#include <sstream>
+#include <iostream>
+namespace OpenGL::Core {
+class Shader {
+ public:
+  // the program ID
+  unsigned int ID;
+  // constructor reads and builds the shader
+  Shader(const char *vertexPath, const char *fragmentPath);
+  // use/activate the shader
+  ~Shader();
+  void use();
+  // utility uniform functions
+  void setBool(const std::string &name, bool value) const;
+  void setInt(const std::string &name, int value) const;
+  void setFloat(const std::string &name, float value) const;
+  void setMat4(const std::string &name, const glm::mat4 &mat) const;
+  void setVec3(const std::string &name, const glm::vec3 &value) const;
+  void setVec3(const std::string &name, float x, float y, float z) const;
+};
+}  // namespace OpenGL::Core
+#endif
