+++ libs/internal_libs/opengl/geometry/src/shape.cpp
@@ -1,0 +1,8 @@
+#include "shape.h"
+#include "textures.h"
+
+void OpenGL::Geometry::Shape::setTexture(std::shared_ptr<OpenGL::Core::Textures> tex) {
+  texture = std::move(tex);
+}
+
+unsigned int OpenGL::Geometry::Shape::getTextureId() const { return texture ? texture->getTexture() : 0; }
