--- libs/internal_libs/shape/include/shape.h
@@ -1,19 +1,0 @@
-#ifndef SHAPE
-#define SHAPE
-#include <memory>
-#include "textures.h"
-
-class Shape {
- public:
-  virtual ~Shape() = default;
-  virtual void init() = 0;
-  virtual void draw(GLuint shaderID, float deltaTime) = 0;
-
-  virtual void setTexture(std::shared_ptr<Textures> tex);
-  virtual unsigned int getTextureId() const;
-
- protected:
-  std::shared_ptr<Textures> texture;
-};
-
-#endif
